third_party/WebKit/LayoutTests/printing/page-height-zero.html

<!DOCTYPE html>
<script src="../resources/js-test.js"></script>
<style>
    @media print {
        /* Need something to trigger multicol layout when entering print mode,
           to trigger the original crash bug (division by zero). */
        #multicol { width: 50%; }
    }
</style>
<div id="multicol" style="-webkit-columns:2;">&nbsp;</div>
<script>
    description("Test that an exception is thrown when attempting to set zero page height");
    if (window.internals) {
        shouldThrow("internals.numberOfPages(100, 0)");
        shouldThrow("internals.pageNumber(document.body, 100, 0)");
    } else {
        testFailed("This test requires window.internals");
    }
</script>
