third_party/blink/web_tests/virtual/layout_ng_block_frag/printing/page-height-zero-expected.txt

Test that an exception is thrown when attempting to set zero page height

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS internals.numberOfPages(100, 0) threw exception TypeError: Failed to execute 'numberOfPages' on 'Internals': Page width and height must be larger than 0..
PASS internals.pageNumber(document.body, 100, 0) threw exception TypeError: Failed to execute 'pageNumber' on 'Internals': Page width and height must be larger than 0..
PASS successfullyParsed is true

TEST COMPLETE

 
